// src/generated/services/DCPO_LISTE_USERService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { DCPO_LISTE_USERRead, DCPO_LISTE_USERWrite } from '../models/DCPO_LISTE_USERModel';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@pa-client/power-code-sdk';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getPowerSdkInstance } from '@pa-client/power-code-sdk';


export class DCPO_LISTE_USERService {
  private static readonly dataSourceName = 'dcpo_liste_user';

  private static readonly client = getPowerSdkInstance(dataSourcesInfo).Data;

  public static async create(record: Omit<DCPO_LISTE_USERWrite, 'ID'>): Promise<IOperationResult<DCPO_LISTE_USERRead>> {
    const result = await DCPO_LISTE_USERService.client.createRecordAsync<Omit<DCPO_LISTE_USERWrite, 'ID'>, DCPO_LISTE_USERRead>(
      DCPO_LISTE_USERService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<DCPO_LISTE_USERWrite, 'ID'>>): Promise<IOperationResult<DCPO_LISTE_USERRead>> {
    const result = await DCPO_LISTE_USERService.client.updateRecordAsync<Partial<Omit<DCPO_LISTE_USERWrite, 'ID'>>, DCPO_LISTE_USERRead>(
      DCPO_LISTE_USERService.dataSourceName,
      id.toString(),
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await DCPO_LISTE_USERService.client.deleteRecordAsync(
      DCPO_LISTE_USERService.dataSourceName,
      id.toString());
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<DCPO_LISTE_USERRead>> {
    const result = await DCPO_LISTE_USERService.client.retrieveRecordAsync<DCPO_LISTE_USERRead>(
      DCPO_LISTE_USERService.dataSourceName,
      id.toString(),
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<DCPO_LISTE_USERRead[]>> {
    const result = await DCPO_LISTE_USERService.client.retrieveMultipleRecordsAsync<DCPO_LISTE_USERRead>(
      DCPO_LISTE_USERService.dataSourceName,
      options
    );
    return result;
  }
/**
* Utility method to get possible values of a referenced entity by its name.
*/
  public static async getReferencedEntity(search: string, referencedEntityKeyNormalized: string): Promise<IOperationResult<unknown>> {
    const params: {search: string} = { search };
    const result = await DCPO_LISTE_USERService.client.executeAsync<{search: string}, unknown>(
      {
        connectorOperation: {
          tableName: DCPO_LISTE_USERService.dataSourceName,
          operationName: 'Get' + referencedEntityKeyNormalized,
          parameters: params
        }
      });
    return result;
  }

}